+++ src/index.js
@@ -2,6 +2,7 @@
 
 import Strophe from './core.js';
 import { $build, $msg, $pres, $iq } from './builder.js';
+import { stx } from './stanza.js';
 
 globalThis.$build = $build;
 globalThis.$iq = $iq;
@@ -9,4 +10,4 @@ globalThis.$msg = $msg;
 globalThis.$pres = $pres;
 globalThis.Strophe = Strophe;
 
-export { $build, $iq, $msg, $pres, Strophe };
+export { $build, $iq, $msg, $pres, Strophe, stx };