+++ src/types/index.d.ts
@@ -3,5 +3,6 @@ import { $iq } from './builder.js';
 import { $msg } from './builder.js';
 import { $pres } from './builder.js';
 import Strophe from './core.js';
-export { $build, $iq, $msg, $pres, Strophe };
+import { stx } from './stanza.js';
+export { $build, $iq, $msg, $pres, Strophe, stx };
 //# sourceMappingURL=index.d.ts.map