+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Type checking via TypeScript and JSDoc typing annotations
   Types definitions are now generated and placed in `./dist/types/`.
 * Remove the deprecated `matchBare` option for `Strophe.Handler`. Use `matchBareFromJid` instead.
+* Add the ability to create stanzas via a tagged template literal (`stx`).
 
 ## Version 1.6.2 - (2023-06-23)